Created new sampling class to unify common methods for all sampling approaches.
